ref/fix: Replace double with float, fix network stability and upgrade NAdam optimizer with corrected momentum estimation

+++ src/main/java/activation/HiddenActivation.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 
 public interface HiddenActivation extends Serializable
 {
-    double activate(double input);
-    double derive(double input);
+    float activate(float input);
+    float derive(float input);
 }
