+++ src/main/java/activation/ReluActivation.java
@@ -6,20 +6,20 @@ public class ReluActivation implements OutputActivation
     @Override
     public void activate(Neuron[] input) {
         int size = input.length;
-        double result;
+        float result;
 
         for(int i = 0; i < size; i++) {
-            result = Math.max(input[i].getValue(), 0.0);
+            result = Math.max(input[i].getValue(), 0.0f);
             input[i].setValue(result);
         }
     }
 
     @Override
-    public void derive(Neuron [] input, double [] output) {
+    public void derive(Neuron [] input, float [] output) {
         int size = input.length;
 
         for(int i = 0; i < size; i++) {
-            input[i].multiplyGrad(input[i].getValue() > 0.0 ? 1.0 : 0.0);
+            input[i].multiplyGrad(input[i].getValue() > 0.0f ? 1.0f : 0.0f);
         }
     }
 }
