+++ src/main/java/execution/Explainer.java
@@ -16,21 +16,21 @@ public class Explainer
     private final NeuralNetwork neuralNetwork;
     private final LayerwiseRelevancePropagation layerwiseRelevancePropagation;
     private final LayerwiseGradientFeatureAttribution layerwiseGradientFeatureAttribution;
-    private double alpha = 1.0;
-    private double beta = 0.0;
+    private float alpha = 1.0f;
+    private float beta = 0.0f;
     private int predictionIndex = 0;
 
-    public Explainer(NeuralNetwork neuralNetwork, double alpha, double beta, int size) {
+    public Explainer(NeuralNetwork neuralNetwork, float alpha, float beta, int size) {
         this.neuralNetwork = neuralNetwork;
         layerwiseRelevancePropagation = new LayerwiseRelevancePropagation(alpha, beta);
         layerwiseGradientFeatureAttribution = new LayerwiseGradientFeatureAttribution(size);
     }
 
-    public double [] performLayerwiseRelevancePropagation(double [] input, NeuralNetwork neuralNetwork) {
+    public float [] performLayerwiseRelevancePropagation(float [] input, NeuralNetwork neuralNetwork) {
        return layerwiseRelevancePropagation.computeLayerwiseRelevancePropagation(input, neuralNetwork);
     }
 
-    public double [] performLayerwiseGradientFeatureAttribution(double [] target, double [] input, NeuralNetwork neuralNetwork) {
+    public float [] performLayerwiseGradientFeatureAttribution(float [] target, float [] input, NeuralNetwork neuralNetwork) {
         return layerwiseGradientFeatureAttribution.computeLayerwiseGradientFeatureAttribution(target, input, neuralNetwork);
     }
