+++ src/main/java/loss/CceLoss.java
@@ -5,28 +5,28 @@ import structure.Neuron;
 public class CceLoss extends AbstractLossFunc
 {
     @Override
-    public double compute(Neuron[] predicted, double [] target) {
+    public float compute(Neuron[] predicted, float [] target) {
         validate(predicted, target);
 
         int size = predicted.length;
-        double cost = 0.0;
+        float cost = 0.0f;
 
         for(int i = 0; i < size; i++) {
-            double value = Math.max(EPSILON, Math.min(1 - EPSILON, predicted[i].getValue()));
-            cost += (target[i] * Math.log(value));
+            float value = Math.max(EPSILON, Math.min(1 - EPSILON, predicted[i].getValue()));
+            cost += (target[i] * (float)Math.log(value));
         }
 
         return -cost;
     }
 
     @Override
-    public void derive(Neuron [] predicted, double [] target) {
+    public void derive(Neuron [] predicted, float [] target) {
         validate(predicted, target);
 
         int size = predicted.length;
 
         for(int i = 0; i < size; i++) {
-            double grad = predicted[i].getValue() - target[i];
+            float grad = predicted[i].getValue() - target[i];
             predicted[i].setGrad(grad);
         }
     }
