+++ src/main/java/regularization/LabelSmoothing.java
@@ -4,20 +4,20 @@ import structure.Neuron;
 
 public class LabelSmoothing
 {
-    private final double epsilon;
+    private final float epsilon;
 
     public LabelSmoothing() {
-        epsilon = 0.0;
+        epsilon = 0.0f;
     }
 
-    public LabelSmoothing(double epsilon) {
+    public LabelSmoothing(float epsilon) {
         this.epsilon = epsilon;
     }
 
-    public void regulate(double [][] target) {
+    public void regulate(float [][] target) {
         for(int i = 0; i < target.length; i++) {
             for(int j = 0; j < target[0].length; j++) {
-                target[i][j] = target[i][j] * (1.0 - epsilon) + epsilon / target[0].length;
+                target[i][j] = target[i][j] * (1.0f - epsilon) + epsilon / target[0].length;
             }
         }
     }
