+++ src/main/java/xai/LayerwiseRelevancePropagation.java
@@ -11,28 +11,28 @@ import java.io.IOException;
 
 public class LayerwiseRelevancePropagation
 {
-    private final double alpha;
-    private final double beta;
+    private final float alpha;
+    private final float beta;
     private int predictionIndex = 0;
 
-    public LayerwiseRelevancePropagation(double alpha, double beta) {
+    public LayerwiseRelevancePropagation(float alpha, float beta) {
         this.alpha = alpha;
         this.beta = beta;
     }
 
-    public double [] computeLayerwiseRelevancePropagation(double [] input, NeuralNetwork neuralNetwork) {
+    public float [] computeLayerwiseRelevancePropagation(float [] input, NeuralNetwork neuralNetwork) {
         Layer[] layer = neuralNetwork.getLayer();
 
-        double [] predict = predict(input, neuralNetwork);
+        float [] predict = predict(input, neuralNetwork);
 
-        double [] relevance = new double [predict.length];
+        float [] relevance = new float [predict.length];
 
         if(relevance.length > 1) {
             predictionIndex = argMax(predict);
-            relevance[predictionIndex] = 1.0;
+            relevance[predictionIndex] = 1.0f;
         }
         else {
-            relevance[0] = 1.0;
+            relevance[0] = 1.0f;
         }
 
         for(int i = layer.length - 1; i >= 1; i--) {
@@ -42,16 +42,16 @@ public class LayerwiseRelevancePropagation
         return compute(layer[0], input, relevance);
     }
 
-    private double [] compute(Layer layer, double [] input, double [] relevance) {
-        double [][] weight = layer.getWeight();
-        double [] newRelevance = new double[input.length];
+    private float [] compute(Layer layer, float [] input, float [] relevance) {
+        float [][] weight = layer.getWeight();
+        float [] newRelevance = new float[input.length];
 
-        double [] positiveSum = new double[relevance.length];
-        double [] negativeSum = new double[relevance.length];
+        float [] positiveSum = new float[relevance.length];
+        float [] negativeSum = new float[relevance.length];
 
         for(int j = 0; j < relevance.length; j++) {
             for(int k = 0; k < newRelevance.length; k++) {
-                double value = input[k]*weight[j][k];
+                float value = input[k]*weight[j][k];
                 positiveSum[j] += Math.max(0, value);
                 negativeSum[j] += Math.min(0, value);
             }
@@ -60,8 +60,8 @@ public class LayerwiseRelevancePropagation
         for(int i = 0; i < newRelevance.length; i++) {
             for(int j = 0; j < relevance.length; j++) {
 
-                double positiveValue = Math.max(0, input[i]*weight[j][i]);
-                double negativeValue = Math.min(0, input[i]*weight[j][i]);
+                float positiveValue = Math.max(0, input[i]*weight[j][i]);
+                float negativeValue = Math.min(0, input[i]*weight[j][i]);
 
                 newRelevance[i] += relevance[j]*((alpha*(positiveValue/positiveSum[j]) - beta*(negativeValue/negativeSum[j])));
             }
@@ -70,12 +70,12 @@ public class LayerwiseRelevancePropagation
         return newRelevance;
     }
 
-    private double [] predict(double [] input, NeuralNetwork neuralNetwork) {
+    private float [] predict(float [] input, NeuralNetwork neuralNetwork) {
         Neuron[] scalarResult = neuralNetwork.forward(input, false);
 
         int size = scalarResult.length;
 
-        double [] result = new double[size];
+        float [] result = new float[size];
 
         for(int i = 0; i < size; i++) {
             result[i] = scalarResult[i].getValue();
@@ -84,7 +84,7 @@ public class LayerwiseRelevancePropagation
         return result;
     }
 
-    private int argMax(double [] output) {
+    private int argMax(float [] output) {
         int bestIndex = 0;
 
         for(int i = 0; i < output.length; i++) {
